# backend/config/swagger.yaml
openapi: 3.0.0
info:
  title: Disaster Relief Application API
  version: 1.0.0
  description: API documentation for the Disaster Relief Application

servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.yourproductiondomain.com/api/v1
    description: Production server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token

  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message

    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]

    Emergency:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: object
          properties:
            type:
              type: string
              example: Point
            coordinates:
              type: array
              items:
                type: number
        status:
          type: string
          enum: [active, resolved]
        severity:
          type: string
          enum: [low, medium, high, critical]

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string

  /emergencies:
    get:
      tags:
        - Emergencies
      summary: Get all emergencies
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - in: query
          name: severity
          schema:
            type: string
          description: Filter by severity level
        - in: query
          name: status
          schema:
            type: string
          description: Filter by status
      responses:
        '200':
          description: List of emergencies
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Emergency'

    post:
      tags:
        - Emergencies
      summary: Create new emergency
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Emergency'
      responses:
        '201':
          description: Emergency created successfully 